requirements.txt
# Core Dependencies
openai>=1.0.0
anthropic>=0.25.0
pyyaml>=6.0
numpy>=1.24.0

# Vector Stores
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Text Processing
tiktoken>=0.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Optional: Local LLM Support
# llama-cpp-python>=0.2.0
# transformers>=4.30.0
# torch>=2.0.0
